+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.12.3
 
+* `Repos`:
+  * `Exposed`:
+    * Add abstract exposed variants of `KeyValue` and `KeyValues` repos
+    * Add new extension `Query#selectPaginated`
+
 ## 0.12.2
 
 * `Versions`: